+++ examples/Visualizer.cpp
@@ -10,6 +10,7 @@
 #include <CLI/Config.hpp>
 
 using namespace candlewick::multibody;
+using pinocchio::visualizers::Vector3;
 using std::chrono::steady_clock;
 namespace fs = std::filesystem;
 
@@ -41,6 +42,7 @@ int main(int argc, char **argv) {
 
   Visualizer visualizer{{window_dims[0], window_dims[1]}, model, geom_model};
   assert(!visualizer.hasExternalData());
+  visualizer.addFrameViz(model.getFrameId("world"), false, Vector3::Ones());
   visualizer.addFrameViz(model.getFrameId("elbow_joint"));
   visualizer.addFrameViz(model.getFrameId("ee_link"));
   pin::Data &data = visualizer.data();