+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.8.0] - 2025-07-25
+- core : Add `LambertMaterial`
+- multibody/Visualizer : remove default world triad
+  * user must add it explicitly
+- multibody/Visualizer : pass scale to addFrameViz()
 
 ## [0.8.0] - 2025-07-25